Removed app settings file, didn't need it
